pyproject.toml
[tool.poetry]
name = "near-expiry-platform"
version = "0.1.0"
description = "AI-powered backend for near-expiry product discount platform"
authors = ["Your Name <your@email.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
pydantic = "^2.6.0"
sqlalchemy = "^2.0.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
langchain = "^0.1.0"
google-cloud-vision = "^3.4.0"
Pillow = "^10.0.0"
imagehash = "^4.3.1"
torch = "^2.0.0"
tensorflow = "^2.15.0"
faiss-cpu = "^1.7.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 